requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
faiss-cpu==1.12.0
sentence-transformers==3.0.0
requests==2.32.3
pydantic==2.10.3